An initial implementation of a map class, hTMap.
This is an initial, not feature-complete implementation of a map class. A map uses one tileset, but can have many individual "tiles" inside it. This required changes to hTSprite and hTRenderer to work with a vairable number of vertices per sprite. Also added a simple hTException class and updated the nlohmann JSON library.