Added a TextBox class that uses an hTMap to draw its shape.
Created a TextBox class that uses a text box-looking map from Tiled. To place it at the top of the screen, I added a ScreenToWorldPosition helper function to hTRenderer. Cleaned up a number of places to be more consistent or correct to handle an object with a dynamic lifetime. Some more thinking about the Render paradigm will be needed soon.