Added hTFont and hTTextBox.
hTFont represents a font in much the same way hTTileset represents a tileset: a texture is provided along with a description of its contents. hTTextBox can be used to render strings of text each of which indexes into hTFont the way hTMap indexes into hTTileset. Very primitive font rendering is now supported. Bounding boxes (with word wrap) and justification are still to come.